Remove the frame pool.
Before we had a fixed size allocator (and so malloc'd frames), it was
a good win. These days, it seems to have no measurable effect on speed
but costs around 1.4MB of memory for Rakudo. More seriously, however,
it was vulnerable when the GC was freeing up closures in a thread it
stole work from (because it used to be blocked), and that thread then
woke up and started executing. This could have been fixed, but given
the current frame cache design not seeming to help a great deal anyway
the easier fix is to remove it.