Fix a few items found by warnings
  * Make some casts explicit
  * Fixup printf formats
  * Remove unused code
  * In MVM_unicode_get_name ensure name is always initialized before use

Note - we can't just turn on -Werror -Wall because of the following:

In src/core/interp.c clang thinks the label "runloop" is unused.  It is
referenced but only inside of a macro.  This is probably a bug in clang.

In the generated file src/strings/unicode.c both CaseFolding_grows_table
and CaseFolding_simple_table are reported as unused.  Do we need these?

When compiling src/jit/emit_posix_x64.c clang complains that the
function emit_win64_callargs in src/jit/emit_x64.dasc is unused.  Since
I'm compiling on MacOS X that makes sense.