emit add, sub, bor, band, bxor with constant if possible
we end up doing one of these two variants 157 times during a
full rakudo compile; sadly it only saves a single mov instruction
if the source and target registers are the same, but it removes
stack access in any case.

i expect the next generation of our jit to do this kind of opt
automatically for pretty much every operation, but that is not now.