Formally bump minimum bytecode version to 5.
The code suggests that versions back to 2 are supported, but it has effectively
been at 4 for a very long time, because there is conditional code relating to
earlier versions. Additionally, the serialization blobs of earlier bytecode
files would be rejected, because they will be using a serialization format
version we now no longer support.