In deserialize_method_cache_lazy(), panic on unimplemented discriminators.
MVM_serialization_write_ref() already panics on encountering an unimplemented
discriminator. deserialize_method_cache_lazy() is designed to ascertain whether
it's possible to defer deserializing the method cache, and takes a very
defensive approach, bailing out for anything it cannot handle. It seems better
to *not* bail out on unimplemented discriminators, and panic instead of
silently deciding "not possible to be lazy", because the silent approach also
hides subtle implementation bugs, and makes refactoring/improvement harder.