Add MVM_oops() function to panic with backtrace
This allows us to panic while leaving a pretty backtrace for
debugging. MVM_panic() is better used for situations in which
we expect memory corruption, but there are still plenty of
situation where some invariant has been invalidated even
without memory corruption. We used to throw exceptions, but
they have the disadvantage that they may be caught and mistaken
for something else.

This only adds MVM_oops() calls in the JIT, as I've most confidence
in placing them over there, but I suspect spesh will also be a
good place for such calls.