Replace a #define with a static inline.
There was a note in the #moarvm backlog that we may be failing to pay
attention to the "allocate in gen2" flag by using the macro too early.
While I can't find any evidence of that, changing to a static inline
should make such a bug less likely to be introduced.