Eliminate 100% core use for async things.
Use async wake-ups instead of an idle handler. For this to work out we
also need a semaphore to avoid a race in sending the first task to the
event loop. GC is also far better handled, using prepare/check handles
so another thread can work-steal the event loop thread for GC purposes
when it is sleeping awaiting an I/O notification etc.