Hopefully fix throwish ops
Before this patch, we used to check the active handler stack
of the interpreter to see if we've entered an exception handler
after executing a 'throwish' opcode in the JIT. However, because
we also have 'goto exceptions', that's not really a very good
check. A better check is to compare the current frame with the JIT
frame.

Because we may resume (yay for delimited continuations) a thrown
exception, even that is not enough, though. So I rigged the exception
handler to store the current jit entry label as an exception resume
label, which we then restore into the resume label. This resume label
is then restored upon resuming a JITted frame, which can actually
happen now we implement more than just die();