Prevent reads of partially deserialized objects.
With lazy deserialization we deserialize objects and STables when they
are first needed. This worked fine in the single-threaded case, but in
a multi-threaded case the mechanism could expose objects and STables
that were only partially deserialized, leading to all manner of odd
behavior and crashes. This patch fixes the problem by making threads
wait if another thread is doing deserialization work on the SC in
question.