Account for inliner's handlers when inlining.
We do this by adding extra entries to the hnadlers table that cover an
entire inlinee and go to the handler in the inliner. This means we can
still find applicable handlers through a linear scan, so the handler
lookup code needs no changes and need not become more complex.