Disable most of the exception optimizations.
This work was fucking sloppy. Not only were the changes to interp.c
bogus, causing random memory overwrites if we ran code in the non-JIT
case, not to mention the memory trashing because category is not a
64-bit number, but the optimizations fail to even check that we know
we're really digging into something with VMException REPR before
rewriting the ops into unchecked memory accesses. All of which led to
SEGVs and crashes of all manner.