Function to return without running exit handlers.
Will be used from a Perl 6 ext-op that fakes up a frame in order to
run constraint checks under it, and then wants to tear it down. If we
do try to run the exit handlers in this case, we've got a frmae that
is not set up for it, and so get all manner of odd behavior.