Implement "free at next safepoint" mechanism.
The fixed size allocator provides a mechanism to free memory, with
the caveat that it is only safe to do so at the next "safe point". A
safe point in MoarVM is currently defined to be a point at which all
threads agree they could now GC, meaning they're certainly not in the
middle of reading a block of memory that another thread is going to
make unreachable to new readers from this point forward.