Add roots and end-of-block stores
Roots are the terminals of the tree. The code generator should
evaluate each root in order. At the end of a block, values are
spilled to memory (that is, until we implement register
allocation). These spills also become roots. This should
ensure that the whole basic block is eventually compiled.