Building .dasc files depends on dynasm version
Because I've patched dynasm to support extended register
addressing, c files generated with an old version of dynasm
are not compatible with the header of the new version. This is
fix by adding a dependency from dynasm-generated c file to the
dynasm scripts, and from the corresponding object file to the
dynasm headers.

At the same time, I've removed the --lua and --has-dynasm options
from the configuration. The --lua option doesn't work well with
the minilua-building rule, and the --has-dynasm option isn't very
useful because our dynasm is now patched (and thus nonstandard).