Implement sp_p6ogetvt_o as expression template
This required the addition of a few new jit ops, notably
do (for a sequence of statements yielding the last), all
(logical and of a series) and any (logical or). and and or
become their corresponding binary operations.

I note that I have not found some way to deal with operand types
(i.e. size) yet, although the sizes are explicit in loads and
stores.