Emit spill should come after IF condition
The condition part of an IF expression is always taken, and may
be the first time a value is computed. Thus, values need to be
spilled there because they may be invalidated elsewhere.

In general the logic of 'invalidation' and 'spilling' is much
to simplistic, because the value may also be computed during
traversal to any other node, and if invalidated the traverser
is not willing to revisit it. (Thus, the value may be lost an
incomputable).