Enforce linear let evaluation order
'let' expressions allow the use of shared subexpressions. If
evaluated in a conditional, the subexpressions may be invalidated,
meaning it is 'lost' for further use. By enforcing 'let' expressions
to be compiled in their order of declaration, we ensure that they
are always available.

In the rare cases that this order is too strict, an optimizer can
potentially reorder statements.