Die when we try to accept() or read() outside an originating thread
Addresses MoarVM #165

libuv doesn't like it when you use a uv_stream_t outside of the
event loop that created it, and since MoarVM has an event loop
per thread, this causes problems.  The problem you see is that your
reads are always zero in length and your accepts never return.

This change is not ideal; I'd rather change MoarVM so that a user
can actually do this (there are some strategies discussed on
https://github.com/MoarVM/MoarVM/issues/165), but this will at least
prevent people from getting a headache trying to figure out why they
aren't getting any data from their sockets.

There are a few other invocations of uv_run in the source code that may
benefit from this as well.