Resolve tiler conflicts by adding new nodes
Tiler conflicts can arise because the expression 'tree' is a DAG,
and consumers of a node may require different tilings. (E.g. one
may require a node to yield a label and another may require a
register).

To resolve these conflicts, we add a new node to the tree, with
the same child nodes but assigned to the new tile. Because tiling
is a bottom-up process, this will not generate a new conflict and
the result must be a tiling without conflicts.