Use fixed-point algorithm to generate rulesets
Rules combine in rulesets because they are equivalent or the
nonterminals they refer to combine. Thus, the key to finding
all rulesets is finding all combinations of nonterminals. This
is achieved by starting the with each nonterminal and iteratively
combining them in a trie. The generated rulesets are then read
from the trie.