Fix tiler table for nodes that don't yield
Some rules (e.g. CARG) can only be placed as children of other
nodes (ARGLIST) and as such never yield either REG or VOID. It
used to be the case that such rules were never added at all,
but that is fixed now. As a side affect, table structure
has been simplified.