Restructure Unicode radix support to fast-path ASCII
This restructures the tests so that the most common/least expensive
cases are checked first: first checks for ASCII 0..9, then ASCII A..Z
and a..z, then the fullwidth letters, and finally a unicode lookup for
Nd characters.

This also gets rid of the gcNd variable meant for shortening the Nd
lookup conditional. It means a longer 'else if' statement, but it also
means no MVM_unicode calls are made until necessary.

These changes undo the speed decreases in NQP and rakudo found when
benchmarking to a much less significant amount.