Cast unsigned short to MVMunit64 to avoid integer overflow
Coverity noticed that in the expression changed here, the variable
`tc->instance->num_user_threads` was an unsigned short, which was then being
implicitly promoted to int and then promoted again to unsigned long.
However, if the result of the expression is sufficiently large, this won't
calculate the correct value anymore since the value overflows.  By casting
`num_user_threads` to MVMunit64 we ensure that the overflow is avoided.