Fix write to null handle on async socket.
This NULL-ing in close_socket was premature, since MVM_free catches this same
handle when the task is processed by the event loop in
close_perform-->close_cb. The problem manifested with code like this:

    react {
      whenever IO::Socket::Async.listen('localhost', 3005) -> $sock {
        start {
          $sock.print("thank you for subscribing to CAT FACTS\n");
          $sock.close;
        }
      }
    }

Occasionally it would succeed, but more often it would segfault on line 320,
with uv_write: at that point the handle was already NULL. A telling
counter-example is this:

    react {
      whenever IO::Socket::Async.listen('localhost', 3005) -> $sock {
        start {
          $sock.print("thank you for subscribing to CAT FACTS\n");
          Promise.in(0).then({ $sock.close });
        }
      }
    }

Where the scheduling of the close task was delayed to the next tick of the
event loop -- after the write task has been dealt with.