Register Locking should be idempotent
Now used AND NOT of flag, which is idempotent to the value itself.
The previous bit trick XOR is not. This caused issues when the same
register was 'used' twice in a

There is another, subtler issue, which is that CALL nodes currently
doubly-use the first register value of CARG, which also subtly
causes wrongness.