Remove more config and generated files in distclean target
As mentioned in GH #139, the `realclean` target doesn't clean everything up.
Actually, it seems that the `distclean` target is the one which is supposed
to clean up the configuration and generated files (at least in MoarVM).
This wasn't happening completely and the current commit removes all files
still needing to be cleaned.

The file `src/jit/emit_posix_x64.c` is force-removed since it might not
exist on non-POSIX systems.  The `config.log` and `config.status` files
within the atomic ops library are also removed, although they really should
be removed as part of the upstream library's `distclean` Makefile target.