Fix data race in frame reference count handling.
Previously, it was possible for a reference to a frame to be acquired
from the body of an object and for it to die before the frame's ref
count was increased. This adds an "acquire frame" operation that will
take care of this issue by ensuring the reference count did not go to
zero already, thus avoiding the race. To ensure the zeroed count will
remain zero after frame death, we arrange to not free the frame's
memory until the next safe point if it was ever referenced by an
object body; the safe point mechanism can still free it right away
in a single-threaded program.