Re-enqueue close stdin request when process is unstarted
Starting a process from Perl 6 using Proc::Async offers no guarantees
that the process has had a chance to run once the start method returns;
it just schedules the process to be started on the async worker thread.
Therefore, it is possible that a request to close standard input may
falsely believe the standard input handle to be closed (by checking
si && si->stdin_handle), when it hasn't actually been set up yet.  This
can result in the child process waiting for input from the parent,
and the parent process waiting for the child to exit.

This change checks if the process has had a chance to run yet when
closing standard input; if it has not, it re-enqueues the task until
that time has come