Make NFG algorithm use Unicode Grapheme Clusters.
As described in Annex #29. We do all of it except the CRLF case, as
enabling that even breaks our ability to parse Perl 6 code (will need
to figure out why). Aside from the CRLF case, though, we now pass all
the Unicode grapheme boundary tests (that is, we get the .chars that
are expected).