Refactor handles for new separator data structure.
Again, no intended semantics changes in this patch. However, we do now
preserve the full set of graphemes in multi-grapheme separators, to
later enable full handling of this case (for now, only the handling of
multiple separators is going to be used, as part of the \r\n grapheme
change).