Store the container and invocation spec flag bits with the bool spec mode.
The boolification spec mode uses 4 bits. Use 3 of the 4 unused bits to store
flags for whether a container spec is present, whether an invocation spec is
present, and whether the hll_owner is non-NULL.

Formerly the two present/absent tests were each stored in an integer.
MVM_serialization_write_str() can store NULL fairly efficiently (For the
hll_owner's name), but as it's fairly common for it to be NULL in the core
setting, we save about 2K by using a flag bit instead for not much extra
complexity.