Remove templates which are compiled incorrectly
The original set of templates was developed with the expressivity
of the template language as our main concern, but this meant that
the JIT compiler had to support a complex set of operations
right from the start. This simplifies that set so that we can
add and test features in relative isolation.