Refactor labelling
Now all basic blocks receive a label equal to their basic block
index. Because we skip the first basic block, the 'zeroth' label is
available for the graph start, and the end-of-graph label is simply
equal to the number of basic blocks (hence 1 greater than the latest
basic block). Individual instructions and data objects are labeled, if
necessary, using an associative array as before. Internal expression
labels are computed during compilation based on the offset.