Let expr tree build consume an iterator
This allows us to safely break a basic block and let the common
graph handling logic deal with difficult cases. This version
can run nqp make test, and, I presume, use OSR correctly.

Also fix a bug (and special case) of the inc_i and friends;
treating a read-write operand simply as a write operand
is much simpler.

There are still plenty of things missing! But at least, and for
the first time, we have a working basis.