start with align=0 as it will grow with attrs
When we compute the memory a CStruct, CPPStruct or CUnion will take, we will pick
the biggest alignment of the struct members. Therefor it makes sense to not start
with a value we might not reach. The previous code made a struct align to not less
than 8 bytes which is problematic for inlined structs. This fixes RT #126675.