Fix use after free
Apparantly the lvalue of tree->nodes[child] was evaluated *before*
the rvalue of assign_tile, and assign_tile may realloc the nodes
array. All this led to ASAN being angry (and rightly so!). This is
fixed by providing a single macro to do it correctly.

Also fix a couple of JIT-related memory leaks, including the
nonclean destruction of graphs that had trees, and the branches
stack that wasn't freed in the register allocator.

nwc10++ for notifying this issue