Add MVM_callsite_copy
When inlining callsites, sometimes two comp units will share a single
callsite in memory.  This makes the issue of who is responsible for
cleaning up after callsites tricky, so instead of incurring extra memory
to track who owns what, we can simply copy the inlined callsite for the
inling comp unit.  This function provides the implementation for
preforming that copy