Allow tiler table generator to specify args
This is actually a complete refactor of the tiler table generator
to support arguments in tiles, to not use a plethora of global
variables, and to compute the optimum rule given a ruleset
more completely. (Which is unfinished, but I think I know how to
complete it). As a result, small changes throughout the tiler and
tile implementations were necessary. eq/lt/gt and friends have also
been reduced to a single cmp tile implementation.