Align bytecode sections to 8 bytes
The bytecode file gets mmapped, and pointers to larger types inside
these sections dereferenced directly, so make sure each section starts
on a maximum alignment boundary.  The internal aligment in each section
is the responsibility of that section's (de)serializers.

This only increases the size of the NQP stage0 bytecode files by about
300 bytes out of a total 2.6MB.