Add illumos build target, minilua-specific flags
Minilua is only a build-time dependency and the only thing we want
from it is to run (correctly). So it doesn't need MoarVM ldlibs,
cflags, or ldflags (and on solaris/sunstudio, it requires another
compiler, too, but I'm not going to fix that in this commit).

A more general flaw of the MoarVM configure system is that it
conflates the OS with its common compiler, which leads to problems
like issue #306 (fixed here).