Add compile-time platform check for JIT support
In issue #324, bus errors are reported when MoarVM is compiled
in a universal binary and runs on a 32 bit platform (OS X).
This adds a compile-time check for the platform, in the assumption
that the compiler will generate different code for each 'instance',
in which case it should disable the JIT support on 32 bit platforms.