fix bug for inlined structures in strutures in nativecall code
An inlined structure loaded from another precompiled compilation unit
errornously reported a wrong alignment, which caused a size increase
and probably a field misalignment in the outer structure.
This was caused by not serializing and deserializing the alignment
information to disk when precompiling a compilation unit.

This patch bumps the serialization format version to 17.