More properly clean up handles on GC.
Including standard handles, which we leaked. Simply adding in a close
and free led to a SEGV, because the event loop has to be run before
it was really save to free the handle. So, this is now done also. It
will need some further revision in the future, but is good enough to
get us a clean sheet with --full-cleanup on most NQP tests.