A more robust strings JIT fix.
All uses of the get_string macro were potentially vulnerable to the
string not having been decoded yet, so this moves the check into the
macro. Also introduce another little function, just to have a clean
way to express that we don't want the string itself, just to be sure
it's decoded.