Don't keep around barely-used bytecode maps.
These essentially doubled the amount of space we used for bytecode
of frames that we actually ran, by keeping around a table noting if
each byte was the start of an instruction. This was used for the low
level backtraces to give an accurate instruction offset and...that's
all!

I can't remember the last time I debugged something using that index.
Worse, it's misleading because it doesn't actually factor in that we
may be in some specialization of the bytecode. So, tossed it for now.
Saves 316KB off NQP and 1088KB off Rakudo running the empty loop
program.