Make frames identifiable using a sequence number
Because frames can now be garbage-collected, they can be moved, and
the identity check used in the JIT will no longer work. Thus, we'd
like to have another cheap way to check if we are currently still
in the same frame.

Sequence numbers are assigned to frames on allocation and assigned
to the thread context on the (re)entry of a frame. Thus, they
provide a per-thread identifier of the position in the call stack.