Mark tc->thread_entry_frame.
Now that frames are collectable, and so can move, this pointer must be
marked so it is kept up to date. Otherwise, it points to junk that may
one day align with the address of an actually executing frame, causing
a worker thread to silently exit and typically resulting in a deadlock
because another thread is blocked on its work.