Fix clearing of "in inter-gen set" flag.
Previously, we toggled it. However, multiple threads can have the
same collectable in their inter-gen set, and may end up considering
it. All the flag does is avoid duplicate additions, so it's safe to
clear it when it could be set. However, before this fix we might have
ended up with it set when it should not have been, which would be
bad.