Insert frame stack -> heap promotion where needed.
We don't allocate any frames on the call stacks yet, and the code to
actually move the frames is not yet implemented. This just inserts the
calls to do the promotion, if needed, in (hopefully all of) the places
where a GC-able object is about to come to reference a frame.