Assign jit entry label prior to invoking
sp_findmeth may be invokish if the spesh slot does not mathc the object
type. We assigned the jit entry label after getting the return value
from MVM_6model_find_spesh, which is 1 if it invoked. This used to be
no problem because FRAME (r12) used to point to the current frame at
all times, but that is no longer true now. Fixed by moving the
assignment prior to the invokish, as in other cases.