Ensure we always produce a heap snapshot.
Force a GC run at the end of heap profiling, so we get an "exit
snapshot". This also means that we'll never spit out empty snapshot
sets if we snapshot a program that runs for a very short time (as we
might when examining the base memory use of NQP/Rakudo).