Stop caching MVMContext on a frame.
We rarely request them anyway, and it should be fairly uncommon to
ask for it more than once. And even then, it's cheap to create, and
while it costs an extra allocation or so we already are using less
space in the nursery for all the other frames (due to losing 8 bytes
thanks to not needing the cache slot), so will very likely come out
ahead even in those cases.