Use frames directly when serializing closures.
Before, we obtained a context ref, just so we could mark the SC on it.
However, now that frames are collectables we can mark an SC directly
on them. This avoids a bunch of allocation and indirection. It also
eliminates the need to have context objects be interned, which was
removed in the previous patch.