this adds a cache for strings that are one codepoint long
in addition, i think it might be nice to also have one
long array of all codepoint values that could be in the
storages of these strings, but that'll add a check when
freeing MVMString objects. Might not be worth it.

I didn't add a mutex because i think it's not needed;
updating the pointer shouldn't break when multiple threads
race it, and since the strings are gc-managed anyway,
things will properly be freed if they get overwritten
by another thread. Also, things that end in the same will
have equal contents, too.