validate indices of param_ ops
we can already identify b0rked frames in the bytecode
validation phase in this aspect. afl-fuzzy was able to
create a crapton of crashing cases based on the access
to param buffers not being bounds-checked at all. Now
instead of paying a little bit for a bounds check on
every single param access, we make validation a bit
smarter up-front.