Distinguish taking all chars from available chars.
The "all chars" indicates we have reached EOF and so want to take all
that is left, flushing the normalization buffer. The "available chars"
indicates that we want to take what is available, but leave anything
in the normalization buffer in place (so an incoming combiner can go
on it, for example).