Move assignment outside boolean expression
The gcc warning `-Wparentheses` noted that one should wrap assignments in
boolean context in parentheses so that they are more obvious within the
boolean expression.  It is clearer to assign the variable in a separate step
and then to run the boolean check.  Making this change also removes the
parentheses warning.  The nqp test suite passes successfully after this
change.