Implement MVM_JIT_IF and MVM_JIT_EITHER via tiling
Because the new register allocator looks at the node type to
construct (merge) live ranges, this is not correct as the node
of the inorder tiles will refer to IF even though they never
refer to any child nodes and do not participate in PHI resolution.
Thus, we need a correct way to distinguish between the inorder
tiles and the postorder tile (which /is/ used in PHI resolution).