Fix a-bit-too-few error in multi cache.
The indexes we store for args are indexes into the args buffer (which
includes names) rather than the callsite flags, meaning that they can
be up to twice as long. Fixes a bug where we could fail to find some
entries when caching callsites with many named args, resulting in the
same entries being put into the cache again and again.