Don't set inheriting process on inherited pipe.
If we spawned process 2, and set its input to a pipe with output from
an already-spawned process 1, then we ended up setting ->process on
the pipe - which formally pointed at process 1 - to point at process
2. This seems somewhat bogus. It ends up with us trying to write to
the C stack because the latter process has no captured I/O and so
plans to store its result there. This in turn leads to the invalid
free - we try to free a pointer pointing into the stack. Removing
this ->process assignment makes the problem go away, and breaks no
spectests. I also can't find any case where it affects previous
behavior on conveynace of exit codes.