Implement naive spilling and splitting
The actual logic of the 'spill' decision needs some work because
it needs to decide whether or not to split a range in the first
place. Also, splits should probably be rangeable (which means
that the split logic may be more complicated).

And it needs the mysterious function 'create_new_live_range'
and 'get_storage_location', which probably aren't really very
complicated but I don't want to think about them right now :-)