Fix CUnion::get_attribute treating inlined attrs as pointers
valgrind reports a Conditional jump or move depends on uninitialised value
when checking if cobj is not NULL. That's because inlined attributes are
not actually pointers but inlined values which may not have been initialized.
The attribute is still there though. Easily visible because the handlers for
the inlined cases don't actually use cobj. Fixed by moving the inlined cases
up and handling them separately.