Fix "Invalid free()" in empty repossessed arrays
If the promised deserialization never happens or a deserialized array is empty
and thus never sees the need to allocate any slots, we could end up free()ing
an uninitialized pointer.
Clean the object in serialization.c as gc_free() functions usually are not
written with the object being re-used in mind.