Implement heap for live ranges
Linear scan must visit live ranges in order of first definition,
and we must be able to add live ranges during traversal.
So we use an array of live range and a heap of indexes into
that array, sorted by first definition.

Also removed some code that could never work.