Fix native callback/GC interaction when embedded.
A while ago, we started to block a thread for GC purposes when doing
a call into native code, so that long-running calls won't prevent
progress of other thread in the VM by preventing their GC. In the
case of callbacks, we then unblocked the thread.

Inline::Perl6 embeds MoarVM. It can therefore create a situation where
a MoarVM native callback is invoked when a call into native code was
not on the stack. This deals with the issue - at least somewhat - by
not trying to unblock a thread that is not blocked, which would lead
to a hang.

This hopefully gets things back to working as well as they did prior
to the nativecall GC blocking fixes. However, it's worth noting that
a situation where code running under Inline::Perl6 starts threads,
and then the Perl 5 code using it keeps the main thread tied up for
a while, would result in the Perl 6 threads also getting blocked.
Thus, Inline::Perl6 may want to explore marking the thread running
the Perl 5 interpreter that embeds MoarVM as GC-blocked to prevent
such a situation.