Stop making trees at frame handler annotations
When we encounter a frame handler annotation, this usually requires
inserting a label, which may invalidate the basic-block property that
code is run without interruption. This patch plugs that hole and allows
rakudo CORE.setting to build again. However, this remains an ad-hoc
way of dealing with such annotations, seeing as from the perspective of
the JIT, theey are really basic block-level attributes rather than
instruction-level attributes.