Do MVM_ASSIGN_REF after block/unblock.
Suppose that we do it before, and the target of the assignment is in
the nursery and has already been seen there once. Thus we don't put
the target into the inter-generational set because it's not in gen2.
Then, when we block/unblock around the lock acquisition, GC triggers
and the target is promoted. At this point the node to add into the
linked list may point to a nursery object, but since we didn't
actually chain the new item into the linked list representing the
queue yet, the GC could never spot it at the point of promotion,
so it would not place the target into the inter-gen set at this
point either. Thus we ended up with a non-inter-gen gen2 object
pointing to a nursery object, leading to all of the usual bad
consequences this may have.