Remove seemingly unneeded and harmfull ldrpath from build
The @prefix@/share/perl6/site/lib directory does not even exist on a normal
installation of rakudo, nqp and moarvm. "@/libdir" does not seem to be needed
either as even with it removed, everything seems to work just fine.
Removing because rpath is prohibited on Fedora and strongly discouraged on
openSUSE.