Fix callstack reset.
In cases where we have recursed heavily, we may end up needing to
flush the entire stack onto the heap. In that case, we need to set
*all* regions to have their ->alloc pointers back to the start of
the region. Otherwise, it can create a situation where we flip into
that region again in the future, and assume that we're at the start
of it - only to end up with memory off the end of it. The resulting
corruption can be highly confusing, looking very much like a GC bug
or an inlining bug (since inlining changes the number of stack frames
allocated, thus hiding the bug).