Mark and detect non-volatile required registers
Some 'definition' tiles (like TC and CU) direct to non-volatile
registers and so need to be handled specially by the register
allocator. This used to be detected by looking at the opcodes of those
tiles, but now this looks at the required register, which is more
flexible and more portable.

This needs to be replicated in some way for the linear scan allocator.