Fix some more bugs with linear_scan
We should not account uses for non-register references. I think there is
still a bug with register releasing.

Because the MVM_JIT_ARCH macro trickery relies on the symbol naem of the
macro, and because we also need to distinguish between possible values,
it's necessary to #define MVM_JIT_ARCH_X64 etc. and later #undef them.

I'm looking into a more regular way to solve this.