Add support for Grapheme_Cluster_Break=Prepend from Unicode 9.0
These marks were added in Unicode 9.0 which occur *before* the
codepoints they combine with. We had assumed that these controlling
characters would not appear before others and only *after*, but with
this change we now pass all the Prepend tests from GraphemeBreakTest.txt
except for a few odd ones ( line Prepend + a space or newline which
shouldn't occur in real life).