Replace ValueRef array buffer with linked list
This reduces the need for the second loop in determine_live_ranges.
As a result of the union proces, the reference lists are merged and
flushed in the 'remaining' live range.

A check in the linear_scan algorithm removes the need for a compaction /
worklist reaping step (although this would also be a valid approach).