Simplifications from making continuations 1-shot.
We no longer need the `in_continuation` flag, which eliminates the
code to set it, the storage (though due to alignment we won't really
save anything) and the code to check it. It was checked every time a
callframe was taken off the stack, which adds up over a program's
lifetime. Furthermore, we can now unconditionally clear up any
continuation tags.