Always release ->work at end of a frame's scope.
This is safe to do unconditionally now that continuations are one-shot
and register references are no more. This brings forward when memory
can be released, and also prevents things from lingering in the
inter-generational set for so long (we now really only ever do that
when the frame is still in dynamic scope - for very long-lived and
heap-promoted frames - since we don't enforce write barriers on
registers).