Eliminate `tc` field in MVMFrame.
We used to use it to track whether a frame is in dynamic scope, but
now the `work` field can be reliably used for that. This shaves a
memory write off every frame setup, but more significantly cuts the
size of every MVMFrame by 4 (32-bit) or 8 (64-bit) bytes, which will
reduce memory pressure.