Tweak full collection criteria in heap profiling.
The recording of heap snapshots will of course use memory, which will
throw off the RSS heuristic and make us a *lot* less likely to ever do
a full collection, distorting the profiles. This is also a bit of a
distortion (to more regular heap profiles being taken), but it's an
improvement. (To do better, we could try tracking RSS before/after
snapshots and excluding that memory from the calculation. Patches
welcome if anyone tries it and finds that a viable appraoch.)