Decode (hopefully) all NFC UTF8 to NFG in UTF8-C8
In the last round of tweaks to UTF8-C8, we fixed some sequences that
would not round-trip properly due to being mis-represented in UTF8.
The fix dealt with those cases, but was a bit too sweeping. UTF8-C8
aims to decode everything that's both valid UTF8 and in NFC as the
UTF8 decoder would, and express everything else as synthetics that
will ensure round-tripping. This fix deals with the issue raised in
MoarVM Issue #482, while not regressing any of the UTF8-C8 roundtrip
tests.