Fix GC in spesh triggered by managed mutex use.
When inlining, the update_mutex of a MVMCompUnit was acquired. This
for some time has been a high-level mutex, made so since these are
reentrant and also cannot lead to GC deadlocks. Unfortunately, spesh
has a "no GC during specialization" constraint, which use of this
mutex violated. Thankfully, the things it used the mutex for were
disjoint from the set of things that need a high-level mutex. So,
split it into two mutexes protecting the different fix, and make the
one that inlining needs to acquire be a low-level mutex which will
never cause entry to the GC.