Resolve ordering of synthetic tiles
Synthetic tiles do not have a 'proper' order number in the list, which
could potentially cause a situation in which a live range was defined
and expired prior to its use (causing it to be overwrritten). By
doubling the tile index and adding a +1/-1 bias for synthetics we use
the additional number space to ensure a correct ordering.