Remove Unicode 1 names and add Unicode Name Aliases
This adds support in the codepointfromname op for all of the official
Unicode name aliases. It removes the old Unicode 1 names in the name
registry and replaces them with <control-xxxx> with the codepoint
number.

There are only 5 codepoints which have changed their functionality
aside from the 400+ Name Aliases added:

Correction:
  * BELL
    * Used to resolve to U+0007 which caused the 🔔 U+1F514 to be
      inaccessible.
    * The following can be used to access U+0007:
      ALERT, or BEL

Unicode 1 Name removed:
  * LINE FEED (LF):
    * The following can be used instead:
      LINE FEED, NEW LINE, END OF LINE, LF, NL or EOL
  * FORM FEED (FF)
    * The following can now be used instead:
      FORM FEED or FF
  * CARRIAGE RETURN (CR)
    * The following can now be used instead:
      CARRIAGE RETURN or CR
  * NEXT LINE (NEL)
    * The following can now be used instead:
      NEXT LINE or NEL# Please enter the commit message for your changes. Lines starting